+++ app/lib/ui/components/buttons/button_outlined.dart
@@ -3,18 +3,18 @@ import 'package:flutter_gen/gen_l10n/ods_flutter_app_localizations.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:ods_flutter/components/button/ods_outlined_button.dart';
 import 'package:ods_flutter/components/sheets_bottom/ods_sheets_bottom.dart';
+import 'package:ods_flutter/guidelines/spacings.dart';
 import 'package:ods_flutter_demo/ui/components/buttons/button_customization.dart';
 import 'package:ods_flutter_demo/ui/main_app_bar.dart';
 
-class ComponentOutlinedButtons extends StatefulWidget {
-  const ComponentOutlinedButtons({super.key});
+class ButtonsOutlined extends StatefulWidget {
+  const ButtonsOutlined({super.key});
 
   @override
-  State<ComponentOutlinedButtons> createState() =>
-      _ComponentOutlinedButtonsState();
+  State<ButtonsOutlined> createState() => _ButtonsOutlinedState();
 }
 
-class _ComponentOutlinedButtonsState extends State<ComponentOutlinedButtons> {
+class _ButtonsOutlinedState extends State<ButtonsOutlined> {
   final _scaffoldKey = GlobalKey<ScaffoldState>();
 
   @override
@@ -26,14 +26,15 @@ class _ComponentOutlinedButtonsState extends State<ComponentOutlinedButtons> {
   Widget build(BuildContext context) {
     return ButtonCustomization(
       child: Scaffold(
-          bottomSheet: OdsSheetsBottom(
-            content: _CustomizationContent(),
-            title: AppLocalizations.of(context)!.componentCustomizeTitle,
-          ),
-          key: _scaffoldKey,
-          appBar: MainAppBar(
-              AppLocalizations.of(context)!.outlinedButtonsVariantTitle),
-          body: _Body()),
+        bottomSheet: OdsSheetsBottom(
+          content: _CustomizationContent(),
+          title: AppLocalizations.of(context)!.componentCustomizeTitle,
+        ),
+        key: _scaffoldKey,
+        appBar: MainAppBar(
+            AppLocalizations.of(context)!.buttonsLowEmphasisVariantTitle),
+        body: SafeArea(child: _Body()),
+      ),
     );
   }
 }
@@ -44,22 +45,18 @@ class _Body extends StatelessWidget {
     final ButtonCustomizationState? customizationState =
         ButtonCustomization.of(context);
 
-    var colorScheme = Theme.of(context).colorScheme;
-    var colorFilter = ColorFilter.mode(colorScheme.primary, BlendMode.srcIn);
-
     return Semantics(
       header: true,
       child: Padding(
-        padding:
-            const EdgeInsets.only(left: 15, top: 25, right: 15, bottom: 15.0),
+        padding: const EdgeInsets.only(
+            left: spacingM, top: spacingL, right: spacingM, bottom: spacingM),
         child: OdsOutlinedButton(
-          title: AppLocalizations.of(context)!.componentButtonsExampleTitle,
+          text: AppLocalizations.of(context)!.componentButtonsExampleTitle,
+          onClick: customizationState?.hasEnabled == true ? () {} : null,
           icon: customizationState?.hasIcon == true
-              ? SvgPicture.asset('assets/ic_about_info.svg',
-                  width: 24, height: 24, colorFilter: colorFilter)
+              ? SvgPicture.asset('assets/ic_profile.svg')
               : null,
-          onPressed: customizationState?.hasEnabled == true ? () {} : null,
-          fullScreenWidth: customizationState?.hasFullScreen ?? false,
+          fullWidth: customizationState?.hasFullScreen ?? false,
         ),
       ),
     );