+++ app/lib/ui/components/components_screen.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:ods_flutter/components/card/ods_small_card.dart';
+import 'package:ods_flutter/guidelines/spacings.dart';
 import 'package:ods_flutter_demo/ui/components/component_entities.dart';
 import 'package:ods_flutter_demo/ui/components/detail_screen.dart';
 import 'package:ods_flutter_demo/ui/utilities/display_image.dart';
@@ -37,30 +38,33 @@ class _ComponentsScreenState extends State<ComponentsScreen> {
     }
 
     return SafeArea(
-      child: GridView.count(
-        crossAxisCount: crossAxisCount,
-        childAspectRatio: childAspectRatio,
-        children: widget.odsComponents.map(
-          (component) {
-            return Column(
-              children: [
-                OdsSmallCard(
-                  title: component.title,
-                  image: displayImage(
-                      component.imageResourceName, double.infinity),
-                  onTap: () {
-                    Get.to(
-                      ComponentDetailScreen(
-                        component: component,
-                      ),
-                      transition: Transition.rightToLeft,
-                    );
-                  },
-                ),
-              ],
-            );
-          },
-        ).toList(),
+      child: Padding(
+        padding: const EdgeInsets.all(spacingS),
+        child: GridView.count(
+          crossAxisCount: crossAxisCount,
+          childAspectRatio: childAspectRatio,
+          children: widget.odsComponents.map(
+            (component) {
+              return Column(
+                children: [
+                  OdsSmallCard(
+                    title: component.title,
+                    image: displayImage(
+                        component.imageResourceName, double.infinity),
+                    onTap: () {
+                      Get.to(
+                        ComponentDetailScreen(
+                          component: component,
+                        ),
+                        transition: Transition.rightToLeft,
+                      );
+                    },
+                  ),
+                ],
+              );
+            },
+          ).toList(),
+        ),
       ),
     );
   }