+++ app/lib/ui/components/dialogs/dialogs.dart
@@ -3,7 +3,7 @@ import 'dart:math';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/ods_flutter_app_localizations.dart';
 import 'package:ods_flutter/components/app_bar/top/ods_top_app_bars.dart';
-import 'package:ods_flutter/components/button/ods_filled_button.dart';
+import 'package:ods_flutter/components/button/ods_button.dart';
 import 'package:ods_flutter/components/dialogs/ods_alert_dialog.dart';
 import 'package:ods_flutter/components/sheets_bottom/ods_sheets_bottom.dart';
 import 'package:ods_flutter/guidelines/spacings.dart';
@@ -60,10 +60,11 @@ class _BodyState extends State<_Body> {
               alignment: WrapAlignment.center,
               spacing: spacingM,
               children: [
-                OdsFilledButton(
-                  title: AppLocalizations.of(context)!
+                OdsButton(
+                  text: AppLocalizations.of(context)!
                       .dialogsVariantExampleOpenButton,
-                  onPressed: () => OdsAlertDialog.openDialog(
+                  style: OdsButtonStyle.functionalPrimary,
+                  onClick: () => OdsAlertDialog.openDialog(
                     context: context,
                     titleText: customizationState?.hasTitle == true
                         ? recipe.title
@@ -87,11 +88,12 @@ class _BodyState extends State<_Body> {
                             : null,
                   ),
                 ),
-                OdsFilledButton(
-                  title: AppLocalizations.of(context)!
+                OdsButton(
+                  text: AppLocalizations.of(context)!
                       .dialogsVariantExampleOpenFullDialogsButton,
-                  fullScreenWidth: false,
-                  onPressed: () => OdsAlertDialog.openFullscreenDialog(
+                  fullWidth: false,
+                  style: OdsButtonStyle.functionalPrimary,
+                  onClick: () => OdsAlertDialog.openFullscreenDialog(
                     context: context,
                     text: recipe.title,
                     body: Padding(