+++ library/lib/components/dialogs/ods_alert_dialog.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:ods_flutter/components/button/ods_filled_button.dart';
+import 'package:ods_flutter/components/button/ods_button.dart';
 import 'package:ods_flutter/guidelines/spacings.dart';
 
 class DismissButtonDescription {
@@ -39,9 +39,10 @@ class OdsAlertDialog extends StatefulWidget {
             child: Text(confirmButtonText),
           ),
           if (dismissButtonDescription != null)
-            OdsFilledButton(
-              title: dismissButtonDescription.dismissButtonText!,
-              onPressed: dismissButtonDescription.onDismissButtonClick,
+            OdsButton(
+              text: dismissButtonDescription.dismissButtonText!,
+              onClick: dismissButtonDescription.onDismissButtonClick,
+              style: OdsButtonStyle.functionalPrimary,
             ),
         ],
       ),